# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "重生之凡人修仙传 API",
    "description": "修仙文字游戏后端接口文档，基于《凡人修仙传》世界观。包含玩家注册登录、角色管理、境界突破、地图探索、闭关修炼、GM管理、系统通知、WebSocket实时推送、玩家数据实时同步等完整功能。",
    "version": "2.2.0",
    "contact": {
      "name": "Shell",
      "email": "1592363624@qq.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发环境"
    },
    {
      "url": "https://xx.52shell.ltd",
      "description": "生产环境"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证"
      }
    },
...
  },
  "paths": {
    "/api/status": {
      "get": {
        "summary": "检查服务器状态",
        "description": "获取服务器和数据库的连接状态，用于健康检查和负载均衡检测",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "服务器正常运行",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "message": {
                      "type": "string",
                      "example": "服务器正常运行，数据库连接成功"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-05-20T12:00:00.000Z"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "server": {
                          "type": "string",
                          "example": "running"
                        },
                        "database": {
                          "type": "string",
                          "example": "connected"
                        },
                        "uptime": {
                          "type": "integer",
                          "description": "服务器运行时间(秒)",
                          "example": 3600
                        },
                        "players_online": {
                          "type": "integer",
                          "description": "在线玩家数",
                          "example": 42
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "数据库连接失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 503
                    },
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "数据库连接失败"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/combat/status": {
      "get": {
        "summary": "获取当前战斗状态",
        "description": "获取当前战斗状态，可通过 battle_id 参数获取指定战斗（从历练进入战斗时使用）",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "battle_id",
            "in": "query",
            "description": "指定战斗ID（可选，用于从历练进入战斗时获取特定战斗）",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "in_battle": { "type": "boolean" },
                    "battle_id": { "type": "string", "nullable": true },
                    "monster": { "type": "object", "nullable": true },
                    "player": { "type": "object", "nullable": true },
                    "round": { "type": "integer" },
                    "turn": { "type": "string", "enum": ["player", "monster"] },
                    "battle_log": { "type": "array" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/encounter": {
      "post": {
        "summary": "遭遇怪物",
        "description": "在当前地图遭遇怪物进入战斗",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "monster_id": { "type": "string", "description": "指定怪物ID（可选）" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "遭遇成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "in_battle": { "type": "boolean" },
                    "battle_id": { "type": "string" },
                    "monster": { "type": "object" },
                    "player": { "type": "object" },
                    "round": { "type": "integer" },
                    "turn": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/attack": {
      "post": {
        "summary": "玩家攻击",
        "description": "在战斗中对怪物发起攻击",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "enum": ["attack", "skill"], "default": "attack" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "攻击结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "in_battle": { "type": "boolean" },
                    "action": { "type": "string" },
                    "damage": { "type": "integer" },
                    "monster_hp": { "type": "string" },
                    "turn": { "type": "string" },
                    "battle_ended": { "type": "boolean" },
                    "result": { "type": "string", "enum": ["victory", "defeat", null] },
                    "exp_gained": { "type": "integer" },
                    "items_gained": { "type": "object" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/monster-turn": {
      "post": {
        "summary": "怪物行动",
        "description": "执行怪物回合行动",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "怪物行动结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "action": { "type": "string" },
                    "damage": { "type": "integer" },
                    "player_hp": { "type": "string" },
                    "turn": { "type": "string" },
                    "round": { "type": "integer" },
                    "battle_ended": { "type": "boolean" },
                    "result": { "type": "string", "enum": ["victory", "defeat", null] }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/flee": {
      "post": {
        "summary": "逃跑",
        "description": "尝试从战斗中逃跑，成功率50%",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "逃跑结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/history": {
      "get": {
        "summary": "获取战斗历史",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "count": { "type": "integer" },
                    "battles": { "type": "array" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/monsters": {
      "get": {
        "summary": "获取当前地图可遭遇怪物列表",
        "description": "获取当前所在地图可遭遇的怪物列表，包含怪物名称、境界和经验奖励",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "map_id": { "type": "integer" },
                    "monsters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "realm": { "type": "string" },
                          "exp": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/skill": {
      "post": {
        "summary": "使用技能",
        "description": "在战斗中使用技能，消耗20点灵力，造成1.5倍伤害",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "skill_index": { "type": "integer", "default": 0, "description": "技能索引" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "技能使用结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "in_battle": { "type": "boolean" },
                    "action": { "type": "string", "example": "skill" },
                    "damage": { "type": "integer" },
                    "mp_used": { "type": "integer", "example": 20 },
                    "monster_hp": { "type": "string" },
                    "player_mp": { "type": "string" },
                    "turn": { "type": "string" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "灵力不足或技能使用失败",
            "content": {
              "application/json": {
                "example": {
                  "code": 400,
                  "message": "灵力不足，需要 20 点灵力"
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/escape": {
      "post": {
        "summary": "逃跑",
        "description": "尝试从战斗中逃跑，成功率50%",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "逃跑结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/combat/stats": {
      "get": {
        "summary": "获取战斗统计",
        "description": "获取玩家的战斗统计数据，包括胜利、失败、逃跑次数和总获得修为",
        "tags": ["Combat"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 200 },
                    "victories": { "type": "integer", "description": "胜利次数" },
                    "defeats": { "type": "integer", "description": "失败次数" },
                    "escapes": { "type": "integer", "description": "逃跑次数" },
                    "total_battles": { "type": "integer", "description": "总战斗次数" },
                    "total_exp": { "type": "integer", "description": "总获得修为" },
                    "win_rate": { "type": "integer", "description": "胜率(%)" },
                    "recent_battles": { "type": "array", "description": "最近战斗记录" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}